Implement `Response` and `Challenge`. Update UNIT-tests and fix bugs. Add incremental build support for UNIT-tests. Shared builds failures are allowed in AppVeyor. Update README.md.

+++ include/ohf/MultipartBody.hpp
@@ -69,11 +69,11 @@ namespace ohf {
 
         std::string boundary();
 
-        Part part(int index);
+        Part part(Uint64 index);
 
         std::vector<Part> parts();
 
-        int size();
+        Uint64 size();
 
         MediaType type();
